Enables list filtering
This PR enables new syntax in the package list header to filter package
lists. This is usefull to filter the installation of packages according
some criteria, such as, the platform specs. As an example, this PR will
allow the new filters: ['gpu'] key in the configuration area of the lst
and the result is that the list will only be part of the stack if every
filter defined in the platform yaml file do not contain the word 'none'